moved components and containers into renderer
