+++ src/req/renderer/components/icons/addBox.ts
