+++ src/req/renderer/components/icons/arrowUpward.ts
