+++ src/req/renderer/components/icons/check.ts
