+++ src/req/renderer/components/icons/chevronRight.ts
