+++ src/req/renderer/components/icons/clear.ts
