+++ src/req/renderer/components/icons/deleteOutline.ts
