+++ src/req/renderer/components/icons/edit.ts
