+++ src/req/renderer/components/icons/filterList.ts
