+++ src/req/renderer/components/icons/firstPage.ts
