+++ src/req/renderer/components/icons/remove.ts
