+++ src/req/renderer/components/icons/search.ts
