+++ src/req/renderer/components/icons/viewColumn.ts
