+++ src/req/renderer/components/tableIcons.ts
