+++ src/req/renderer/components/tabs.ts
