+++ src/req/renderer/containers/fullWidthTabs.tsx
